% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_functions.R
\name{changes.list}
\alias{changes.list}
\title{Enumerate Changes to a ResourceRecordSet collection.}
\usage{
changes.list(project, managedZone, maxResults = NULL, pageToken = NULL,
  sortBy = NULL, sortOrder = NULL)
}
\arguments{
\item{project}{Identifies the project addressed by this request}

\item{managedZone}{Identifies the managed zone addressed by this request}

\item{maxResults}{Optional}

\item{pageToken}{Optional}

\item{sortBy}{Sorting criterion}

\item{sortOrder}{Sorting order direction: 'ascending' or 'descending'}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/ndev.clouddns.readonly
\item https://www.googleapis.com/auth/ndev.clouddns.readwrite
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/ndev.clouddns.readonly, https://www.googleapis.com/auth/ndev.clouddns.readwrite)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/cloud-dns}{Google Documentation}
}

