% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ChangesListResponse}
\alias{ChangesListResponse}
\title{ChangesListResponse Object}
\usage{
ChangesListResponse(changes = NULL, nextPageToken = NULL)
}
\arguments{
\item{changes}{The requested changes}

\item{nextPageToken}{The presence of this field indicates that there exist more results following your last page of results in pagination order}
}
\value{
ChangesListResponse object
}
\description{
ChangesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a request to enumerate Changes to a ResourceRecordSets collection.
}

