% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{Project}
\alias{Project}
\title{Project Object}
\usage{
Project(id = NULL, number = NULL, quota = NULL)
}
\arguments{
\item{id}{User assigned unique identifier for the resource (output only)}

\item{number}{Unique numeric identifier for the resource; defined by the server (output only)}

\item{quota}{Quotas assigned to this project (output only)}
}
\value{
Project object
}
\description{
Project Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A project resource. The project is a top level container for resources including Cloud DNS ManagedZones. Projects can be created only in the APIs console.
}
