% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{Quota}
\alias{Quota}
\title{Quota Object}
\usage{
Quota(managedZones = NULL, resourceRecordsPerRrset = NULL,
  rrsetAdditionsPerChange = NULL, rrsetDeletionsPerChange = NULL,
  rrsetsPerManagedZone = NULL, totalRrdataSizePerChange = NULL)
}
\arguments{
\item{managedZones}{Maximum allowed number of managed zones in the project}

\item{resourceRecordsPerRrset}{Maximum allowed number of ResourceRecords per ResourceRecordSet}

\item{rrsetAdditionsPerChange}{Maximum allowed number of ResourceRecordSets to add per ChangesCreateRequest}

\item{rrsetDeletionsPerChange}{Maximum allowed number of ResourceRecordSets to delete per ChangesCreateRequest}

\item{rrsetsPerManagedZone}{Maximum allowed number of ResourceRecordSets per zone in the project}

\item{totalRrdataSizePerChange}{Maximum allowed size for total rrdata in one ChangesCreateRequest in bytes}
}
\value{
Quota object
}
\description{
Quota Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Limits associated with a Project.
}
