% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ResourceRecordSet}
\alias{ResourceRecordSet}
\title{ResourceRecordSet Object}
\usage{
ResourceRecordSet(name = NULL, rrdatas = NULL, ttl = NULL, type = NULL)
}
\arguments{
\item{name}{For example, www}

\item{rrdatas}{As defined in RFC 1035 (section 5) and RFC 1034 (section 3}

\item{ttl}{Number of seconds that this ResourceRecordSet can be cached by resolvers}

\item{type}{The identifier of a supported record type, for example, A, AAAA, MX, TXT, and so on}
}
\value{
ResourceRecordSet object
}
\description{
ResourceRecordSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A unit of data that will be returned by the DNS servers.
}

